Refactor announcement hooks import paths for consistency
- Updated import paths for announcement-related hooks in multiple components to use relative paths instead of absolute paths.
- This change enhances code organization and maintainability by standardizing the import structure across the announcement components.
